diagops: hom kernel connector
